+++ epictrack-web/src/components/project/ProjectList.tsx
@@ -15,7 +15,7 @@ import { ProjectDialog } from "./Dialog";
 import { showNotification } from "components/shared/notificationProvider";
 import { useCachedState } from "hooks/useCachedFilters";
 import { ColumnFilter } from "components/shared/MasterTrackTable/type";
-import { exportToCsv } from "utils/exportUtils";
+import { exportToCsv } from "components/shared/MasterTrackTable/utils";
 import { FileDownload } from "@mui/icons-material";
 
 const projectsListingFiltersCacheKey = "projects-listing-filters";
