+++ epictrack-web/src/components/shared/MasterTrackTable/utils.ts
@@ -1,3 +1,8 @@
+import { json2csv } from "json-2-csv";
+import { dateUtils } from "utils";
+import { MRT_TableInstance } from "material-react-table";
+import { MRT_RowData } from "material-react-table";
+
 export const BLANK_OPTION = "(Blanks)";
 export function getSelectFilterOptions<T>(
   data: T[],
@@ -60,3 +65,37 @@ export const rowsPerPageOptions = (dataSize = 10) => {
 
   return defaultOptions;
 };
+
+interface ExportToCsvOptions<T extends MRT_RowData> {
+  table: MRT_TableInstance<T>;
+  downloadDate: string | null;
+  filenamePrefix: string;
+}
+
+export async function exportToCsv<T extends MRT_RowData>({
+  table,
+  downloadDate,
+  filenamePrefix,
+}: ExportToCsvOptions<T>) {
+  const filteredResult = table
+    .getFilteredRowModel()
+    .flatRows.map((p) => p.original);
+  const columns = table
+    .getVisibleFlatColumns()
+    .map((p) => p.columnDef.id?.toString());
+  const csv = await json2csv(filteredResult, {
+    emptyFieldValue: "",
+    keys: columns as string[],
+  });
+  const url = window.URL.createObjectURL(new Blob([csv as any]));
+  const link = document.createElement("a");
+  link.href = url;
+  link.setAttribute(
+    "download",
+    `${filenamePrefix}-${dateUtils.formatDate(
+      downloadDate ? downloadDate : new Date().toISOString()
+    )}.csv`
+  );
+  document.body.appendChild(link);
+  link.click();
+}
