+++ epictrack-web/src/components/work/workStaff/WorkStaffList.tsx
@@ -7,7 +7,7 @@ import MasterTrackTable from "../../shared/MasterTrackTable";
 import { useCachedState } from "hooks/useCachedFilters";
 import { ColumnFilter } from "components/shared/MasterTrackTable/type";
 import { ETPageContainer } from "components/shared";
-import { exportToCsv } from "utils/exportUtils";
+import { exportToCsv } from "components/shared/MasterTrackTable/utils";
 import { FileDownload } from "@mui/icons-material";
 
 const workStaffListColumnFiltersCacheKey = "work-staff-listing-column-filters";
